kloud: initial refactoring to external package
