+++ go/src/github.com/koding/kloud/api/digitalocean/digitalocean.go
