+++ go/src/github.com/koding/kloud/api/digitalocean/droplet.go
