+++ go/src/github.com/koding/kloud/api/digitalocean/image.go
