+++ go/src/github.com/koding/kloud/api/digitalocean/key.go
