+++ go/src/github.com/koding/kloud/api/digitalocean/types.go
