+++ go/src/github.com/koding/kloud/api/openstack/flavor.go
