+++ go/src/github.com/koding/kloud/api/openstack/image.go
