+++ go/src/github.com/koding/kloud/api/openstack/key.go
