+++ go/src/github.com/koding/kloud/idlock/README.md
