+++ go/src/github.com/koding/kloud/idlock/idlock.go
