+++ go/src/github.com/koding/kloud/kloudctl/command/destroy.go
