+++ go/src/github.com/koding/kloud/kloudctl/command/ping.go
