+++ go/src/github.com/koding/kloud/kloudctl/command/restart.go
