+++ go/src/github.com/koding/kloud/kloudctl/command/stop.go
