+++ go/src/github.com/koding/kloud/packer/packer.go
