+++ go/src/github.com/koding/kloud/provider/digitalocean/_cache.go
