+++ go/src/github.com/koding/kloud/provider/openstack/state.go
@@ -1,8 +1,9 @@
 package openstack
 
 import (
-	"koding/kites/kloud/kloud/machinestate"
 	"strings"
+
+	"github.com/koding/kloud/machinestate"
 )
 
 // statusToState converts a rackspacke status to a sensible machinestate.State
