--- go/src/koding/kites/kloud/kloud/deploy.go
@@ -1,172 +1,0 @@
-package kloud
-
-import (
-	"fmt"
-	"koding/kites/kloud/kloud/protocol"
-	"koding/kites/kloud/sshutil"
-	"path/filepath"
-	"strings"
-	"time"
-
-	klientprotocol "koding/kites/klient/protocol"
-
-	kiteprotocol "github.com/koding/kite/protocol"
-	"github.com/nu7hatch/gouuid"
-	"github.com/pkg/sftp"
-)
-
-func (k *Kloud) DeployFunc(username, hostname, ipAddress string) (*protocol.DeployArtifact, error) {
-	log := func(msg string) {
-		k.Log.Info("%s ==> %s", username, msg)
-	}
-
-	sshAddress := ipAddress + ":22"
-	sshConfig, err := sshutil.SshConfig(protocol.PrivateKey)
-	if err != nil {
-		return nil, err
-	}
-
-	log("Connecting to SSH: " + sshAddress)
-	client, err := sshutil.ConnectSSH(sshAddress, sshConfig)
-	if err != nil {
-		return nil, err
-	}
-	defer client.Close()
-
-	sftp, err := sftp.NewClient(client.Client)
-	if err != nil {
-		return nil, err
-	}
-
-	log("Creating a kite.key directory")
-	err = sftp.Mkdir("/etc/kite")
-	if err != nil {
-		return nil, err
-	}
-
-	tknID, err := uuid.NewV4()
-	if err != nil {
-		return nil, NewError(ErrSignGenerateToken)
-	}
-
-	log("Creating a key with kontrolURL: " + k.KontrolURL)
-	kiteKey, err := k.createKey(username, tknID.String())
-	if err != nil {
-		return nil, err
-	}
-
-	remoteFile, err := sftp.Create("/etc/kite/kite.key")
-	if err != nil {
-		return nil, err
-	}
-
-	log("Copying kite.key to remote machine")
-	_, err = remoteFile.Write([]byte(kiteKey))
-	if err != nil {
-		return nil, err
-	}
-
-	log("Fetching latest klient.deb binary")
-	latestDeb, err := k.Bucket.Latest()
-	if err != nil {
-		return nil, err
-	}
-
-	// splitted => [klient 0.0.1 environment arch.deb]
-	splitted := strings.Split(latestDeb, "_")
-	if len(splitted) != 4 {
-		// should be a valid deb
-		return nil, fmt.Errorf("invalid deb file: %v", latestDeb)
-	}
-	kiteVersion, kiteEnv := splitted[1], splitted[2]
-
-	// signedURL allows us to have public access for a limited time frame
-	signedUrl := k.Bucket.SignedURL(latestDeb, time.Now().Add(time.Minute*3))
-
-	log("Downloading '" + filepath.Base(latestDeb) + "' to /tmp inside the machine")
-	out, err := client.StartCommand(fmt.Sprintf("wget -O /tmp/klient-latest.deb '%s'", signedUrl))
-	if err != nil {
-		fmt.Println("out", out)
-		return nil, err
-	}
-
-	log("Installing klient deb on the machine")
-	out, err = client.StartCommand("dpkg -i /tmp/klient-latest.deb")
-	if err != nil {
-		fmt.Println("out", out)
-		return nil, err
-	}
-
-	log("Removing leftover klient deb from the machine")
-	out, err = client.StartCommand("rm -f /tmp/klient-latest.deb")
-	if err != nil {
-		fmt.Println("out", out)
-		return nil, err
-	}
-
-	log("Restarting klient with kite.key")
-	out, err = client.StartCommand("service klient restart")
-	if err != nil {
-		return nil, err
-	}
-
-	// arslan/public-host/klient/0.0.1/unknown/testkloud-1401755272229370184-0/393ff626-8fa5-4713-648c-4a51604f98c6
-	query := kiteprotocol.Kite{
-		Username:    username,       // kite.key is signed for this user
-		ID:          tknID.String(), // id is generated by ourself
-		Hostname:    hostname,       // hostname is the dropletName
-		Name:        klientprotocol.Name,
-		Environment: kiteEnv,
-		Region:      klientprotocol.Region,
-		Version:     kiteVersion,
-	}
-
-	k.Log.Info("Connecting to remote Klient instance")
-	klient, err := k.Klient(query.String())
-	if err != nil {
-		k.Log.Warning("Connecting to remote Klient instance err: %s", err)
-	} else {
-		k.Log.Info("Sending a ping message")
-		if err := klient.Ping(); err != nil {
-			k.Log.Warning("Sending a ping message err:", err)
-		}
-	}
-
-	return &protocol.DeployArtifact{
-		KiteQuery: query.String(),
-	}, nil
-}
-
-// changeHostname is used to change the remote machines hostname by modifying
-// their /etc/host and /etc/hostname files.
-func changeHostname() error {
-	// hostFile, err := client.Create("/etc/hosts")
-	// if err != nil {
-	// 	return nil, err
-	// }
-	//
-	// if err := t.Execute(hostFile, hostname); err != nil {
-	// 	return nil, err
-	// }
-	//
-	// hostnameFile, err := client.Create("/etc/hostname")
-	// if err != nil {
-	// 	return nil, err
-	// }
-	//
-	// _, err = hostnameFile.Write([]byte(hostname))
-	// if err != nil {
-	// 	return nil, err
-	// }
-
-	// if err := client.StartCommand(fmt.Sprintf("hostname %s", hostname)); err != nil {
-	// 	return nil, err
-	// }
-	// out, err = client.StartCommand("service networking restart")
-	// if err != nil {
-	// 	return nil, err
-	// }
-	//
-
-	return nil
-}
